Refactor: 프로필 스와이퍼·연락처·소셜 링크를 전역 공유 컴포넌트로 분리
- @side 전용이던 contactInfo·profileSwiper를 components/로 이동
- @side에 인라인이던 깃허브·블로그·링크드인 링크 카드를 socialLinks 컴포넌트로 추출
- 모바일 히어로와 @side가 함께 사용하기 위한 사전 분리
